+++ src/ApiPlatform/Metadata/ExtrasPropertyNameCollectionFactory.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace CoolMS\EntityBundle\ApiPlatform\Metadata;
+namespace CoolMS\Entity\Bundle\ApiPlatform\Metadata;
 
 use ApiPlatform\Metadata\Exception\ResourceClassNotFoundException;
 use ApiPlatform\Metadata\Property\Factory\PropertyNameCollectionFactoryInterface;