+++ src/ApiPlatform/Resource/EntityFieldsResource.php
@@ -2,13 +2,13 @@
 
 declare(strict_types=1);
 
-namespace CoolMS\EntityBundle\ApiPlatform\Resource;
+namespace CoolMS\Entity\Bundle\ApiPlatform\Resource;
 
 use ApiPlatform\Metadata\ApiResource;
 use ApiPlatform\Metadata\Get;
 use CoolMS\Entity\Field\FieldDescriptor;
 use CoolMS\Entity\Filter\VirtualFieldDescriptor;
-use CoolMS\EntityBundle\ApiPlatform\Resource\Provider\EntityFieldsProvider;
+use CoolMS\Entity\Bundle\ApiPlatform\Resource\Provider\EntityFieldsProvider;
 
 /**
  * Read-only resource exposing the filterable /